challenges/jollyroger/requirements.txt
crcmod
pyelftools
